Restructures so it resembles the structure we have else where
