+++ src/main/js/hooks/FormValidation.js
@@ -9,7 +9,7 @@ function useFormValidation(initialState, validate, submitHandler) {
         if (isSubmitting) {
             const noErrors = Object.keys(errors).length === 0;
             if (noErrors) {
-                await submitHandler();
+                await submitHandler(values);
                 setSubmitting(false);
             } else {
                 setSubmitting(false);
